Close three gaps in the workflow contract helpers
Three review findings from the 11:51 round that I had missed.

**A cache step with no paths passed every ownership assertion.**
`cache_paths` defaulted a missing `with.path` to the empty string, so a
restore or save step that mounted nothing yielded an empty list and every
ownership, uv-directory, and forbidden-path check over it held
vacuously. It now requires the input to be present and to yield at least
one path.

**The compiler-cache report could end before the work it measures.**
The reset already sits immediately after the wrapper is installed, which
bounds the window below. This adds the upper bound: the report must come
after the last step that can invoke the compiler. Some prefixes also
match steps that compile nothing, `Check out repository` among them,
which is harmless because only the last index is used.

**The public helpers lacked NumPy sections.** Parameters and Returns
added throughout, per the repository convention for public APIs.

Those docstrings pushed the module past the 400-line limit, so it splits
along its natural seam: `ci_workflows.py` reads workflows and recognizes
cache steps, `ci_runners.py` holds the manifests that say what is
intended. The manifest module re-exports the readers, so every contract
test still imports from one place.
